% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello45-package.R
\name{hello45}
\alias{hello45}
\title{It does some stuff}
\description{
It does some stuff
}
\section{Support}{


The development repository for the **hello45** package is found at
\url{https://github.com/noamross/hello45}.  Please file
bug reports or other feedback at \url{https://github.com/noamross/hello45/issues}
}

\author{
Noam Ross \email{noam.ross@gmail.com}
}
\keyword{package}
